reorganize CLI and dev utils
