+++ server/setup.cfg
@@ -19,7 +19,7 @@ install_requires =
     gene-normalizer >= 0.1.27
     biocommons.seqrepo
     uta-tools >= 0.0.17
-    fusor >= 0.0.18
+    fusor >= 0.0.19
     sqlparse >= 0.4.2
     websockets >= 9.1
     uvloop
@@ -45,9 +45,8 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    build_fusion_utils = curation.cli:utils
+    fusion_devtools = curation.cli:devtools
     serve_fusion_curation = curation.cli:serve
-    build_client_types = curation.utils.build_client_types:build_client_types_file
 
 [tool:pytest]
 addopts = --ignore setup.py --ignore codebuild --doctest-modules --disable-warnings --cov-report term-missing --cov .
